Ensure load metadata from homepage
