+++ src/pages/stats/StatsPage.tsx
@@ -2,74 +2,22 @@ import { Grid } from '@material-ui/core';
 import React from 'react';
 import FilterSection from '../../components/FilterSection';
 import GraphCard from '../../components/GraphCard';
-import useCharacterBonus from '../../components/hooks/useCharacterBonus';
-import useCharactersMetadata from '../../components/hooks/useCharactersMetadata';
 import useFilteredData from '../../components/hooks/useFilteredData';
-import useFragmentMetadata from '../../components/hooks/useFragmentsMetadata';
-import useStagesMetadata from '../../components/hooks/useStagesMetadata';
 import RankTableCard from '../../components/RankTableCard';
 import { FilterContext } from '../../context/filterContext';
-import { MetadataContext } from '../../context/metadataContext';
-import filters, { createFragmentFilters } from '../../filters';
+import { SelectedCharacterBonus } from '../../types';
 
-const StatsPage = () => {
+const StatsPage = ({
+  data,
+  isLoading,
+}: {
+  data: SelectedCharacterBonus[];
+  isLoading: boolean;
+}) => {
   //Contexts
-  const { dispatch } = React.useContext(MetadataContext);
-  const { dispatch: filterDispatch, state: filterContext } =
-    React.useContext(FilterContext);
-
-  //Fetch data
-  const { data: stageMetadata, isLoading: isStageMetadataLoading } =
-    useStagesMetadata();
-  const { data: characterMetadata, isLoading: isCharacterMetadataLoading } =
-    useCharactersMetadata();
-  const { data: fragmentMetadata, isLoading: isFragmentMetadataLoading } =
-    useFragmentMetadata();
-  const { data, isLoading: isDataLoading } = useCharacterBonus();
+  const { state: filterContext } = React.useContext(FilterContext);
 
   const filteredData = useFilteredData(data, filterContext.filters);
-
-  React.useEffect(() => {
-    dispatch({ type: 'calculateLastMonth', data });
-  }, [dispatch, data]);
-
-  React.useEffect(() => {
-    if (!isStageMetadataLoading) {
-      dispatch({ type: 'setStages', stages: stageMetadata });
-    }
-
-    if (!isCharacterMetadataLoading) {
-      dispatch({ type: 'setCharacters', characters: characterMetadata });
-    }
-
-    if (!isFragmentMetadataLoading && fragmentMetadata) {
-      dispatch({ type: 'setFragments', fragments: fragmentMetadata });
-    }
-  }, [
-    dispatch,
-    characterMetadata,
-    isCharacterMetadataLoading,
-    isStageMetadataLoading,
-    isFragmentMetadataLoading,
-    stageMetadata,
-    fragmentMetadata,
-  ]);
-
-  React.useEffect(() => {
-    const fragmentFilters = createFragmentFilters(fragmentMetadata || []);
-
-    filterDispatch({
-      type: 'add',
-      filters: [...filters, ...fragmentFilters],
-    });
-  }, [filterDispatch, fragmentMetadata]);
-
-  const isLoading =
-    isStageMetadataLoading &&
-    isCharacterMetadataLoading &&
-    isFragmentMetadataLoading &&
-    isDataLoading;
-
   return (
     <>
       <Grid container spacing={3}>
